backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
# Export dependencies
weasyprint==60.0
python-docx==0.8.11
markdownify==0.11.6
